+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.21.0] - 2024-12-18
+
 ### Changed
 
 - The build now fails early if known problematic Python and pip-related env vars have been set by the user or earlier buildpacks. ([#308](https://github.com/heroku/buildpacks-python/pull/308))
@@ -264,7 +266,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation. ([#3](https://github.com/heroku/buildpacks-python/pull/3))
 
-[unreleased]: https://github.com/heroku/buildpacks-python/compare/v0.20.1...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-python/compare/v0.21.0...HEAD
+[0.21.0]: https://github.com/heroku/buildpacks-python/compare/v0.20.1...v0.21.0
 [0.20.1]: https://github.com/heroku/buildpacks-python/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/heroku/buildpacks-python/compare/v0.19.2...v0.20.0
 [0.19.2]: https://github.com/heroku/buildpacks-python/compare/v0.19.1...v0.19.2